Added get order to the DataSource

+++ goods-demo-ui/src/datasource/db/DBDatasource.h
@@ -96,7 +96,7 @@ public:
 	OrderPtr registerOrder(const CreateOrderParams& orderParams) override;
 	uint64_t getNextOrderNumber() override;
 	OrdersList allOrdersByCustomer(const std::string& customerEmail) override;
-	OrderPtr getOrder(uint32_t number, const std::string& customerEmail) override;
+	OrderPtr getOrder(uint64_t number, const std::string& customerEmail) override;
 	~DBDataSource();
 };
